backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# Database
supabase==2.10.0
psycopg2-binary==2.9.11

# Authentication
pyjwt==2.10.0
cryptography==44.0.0
bcrypt==4.0.1  # Compatible with passlib 1.7.4
passlib[bcrypt]==1.7.4

# AI/ML
huggingface-hub==0.26.2
sentence-transformers==3.3.1
torch>=2.6.0
transformers==4.46.3
numpy>=1.24.0,<2.0.0

# APIs
httpx>=0.26,<0.28
aiohttp==3.11.7

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.4

# Utilities
pydantic==2.10.1
pydantic-settings==2.6.1
email-validator==2.3.0
deep-translator==1.11.4
openai>=1.0.0  # For OpenRouter API (Gemini integration)

# CORS
python-jose[cryptography]==3.3.0

# Testing (optional for now)
pytest==8.3.3
pytest-asyncio==0.24.0
